+++ CHANGELOG.md
@@ -17,6 +17,53 @@
 
 * we rename components.css to warp.css as it can be used without using any components
 
+# [1.3.0-next.1](https://github.com/warp-ds/css/compare/v1.2.1-next.1...v1.3.0-next.1) (2023-11-01)
+
+
+### Features
+
+* **attention:** add close button classes ([#80](https://github.com/warp-ds/css/issues/80)) ([5923b22](https://github.com/warp-ds/css/commit/5923b22368493dac0a9a5b0f0ee358a10aad2b51))
+
+## [1.2.1-next.1](https://github.com/warp-ds/css/compare/v1.2.0...v1.2.1-next.1) (2023-10-27)
+
+
+### Bug Fixes
+
+* fix modal sorting problem in pill button ([#77](https://github.com/warp-ds/css/issues/77)) ([736429f](https://github.com/warp-ds/css/commit/736429f257741bdb167fb6df99d491661bd3119b))
+
+# [1.2.0](https://github.com/warp-ds/css/compare/v1.1.2...v1.2.0) (2023-10-25)
+
+
+### Bug Fixes
+
+* adjust chevron animation ([#72](https://github.com/warp-ds/css/issues/72)) ([51e0cd1](https://github.com/warp-ds/css/commit/51e0cd1d26f52a627c75e32609c61eecf19a3360))
+
+
+### Features
+
+* **component-classes:** add 'highlight' variant to attention classes ([#74](https://github.com/warp-ds/css/issues/74)) ([071a624](https://github.com/warp-ds/css/commit/071a6244cbef75b9880496a7b8c1e42a114e5b92))
+
+# [1.2.0-next.1](https://github.com/warp-ds/css/compare/v1.1.3-next.1...v1.2.0-next.1) (2023-10-25)
+
+
+### Features
+
+* **component-classes:** add 'highlight' variant to attention classes ([#74](https://github.com/warp-ds/css/issues/74)) ([071a624](https://github.com/warp-ds/css/commit/071a6244cbef75b9880496a7b8c1e42a114e5b92))
+
+## [1.1.3-next.1](https://github.com/warp-ds/css/compare/v1.1.2...v1.1.3-next.1) (2023-10-23)
+
+
+### Bug Fixes
+
+* adjust chevron animation ([#72](https://github.com/warp-ds/css/issues/72)) ([51e0cd1](https://github.com/warp-ds/css/commit/51e0cd1d26f52a627c75e32609c61eecf19a3360))
+
+## [1.1.2](https://github.com/warp-ds/css/compare/v1.1.1...v1.1.2) (2023-10-20)
+
+
+### Bug Fixes
+
+* **button:** add text-center to buttonAsLink ([#67](https://github.com/warp-ds/css/issues/67)) ([a67b6a4](https://github.com/warp-ds/css/commit/a67b6a416e4cb385daf56c811db181e30dfeb591))
+
 ## [1.1.2-next.1](https://github.com/warp-ds/css/compare/v1.1.1...v1.1.2-next.1) (2023-10-13)